% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cel_to_fah.R
\name{cel_to_fah}
\alias{cel_to_fah}
\title{Convert Celsius degrees to Fahrenheit}
\usage{
cel_to_fah(x)
}
\arguments{
\item{x}{of Celsius degrees is a numerical (double)}
}
\value{
Fahrenheit degrees as numerical (double)
}
\description{
cel_to_fah is a function to convert Celsius degrees to Fahrenhit.
Insert temperature in Celsius degrees to get the value in Fahrenheit degrees.
For further information access:
https://github.com/marcelofukuoka/conversorR
}
\examples{
cel_to_fah(30)
}
