% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cel_to_kel.R
\name{cel_to_kel}
\alias{cel_to_kel}
\title{Convert Celsius degrees to Kelvin(K)}
\usage{
cel_to_kel(x)
}
\arguments{
\item{x}{of Celsius degrees is a numerical (double)}
}
\value{
Kelvin(K) as numerical (double)
}
\description{
cel_to_kel is a function to convert Celsius degrees to Kelvin(K)
Insert temperature in Celsius degrees to get the value in Kelvin(K).
For further information access:
https://github.com/marcelofukuoka/conversorR
}
\examples{
cel_to_Kel(26.85)
}
