% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fah_to_cel.R
\name{fah_to_cel}
\alias{fah_to_cel}
\title{Convert Fahrenheit degrees to Celsius}
\usage{
fah_to_cel(x)
}
\arguments{
\item{x}{of Fahrenheit degrees is a numerical (double)}
}
\value{
Celsius degrees as numerical (double)
}
\description{
fah_to_cel is a function to convert Fahrenheit degrees to Celsius.
Insert temperature in Fahrenheit degrees to get the value in Celsius degrees.
For further information access:
https://github.com/marcelofukuoka/conversorR
}
\examples{
fah_to_cel(86)
}
