% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fah_to_kel.R
\name{fah_to_kel}
\alias{fah_to_kel}
\title{Convert Fahrenheit degrees to Kelvin(K)}
\usage{
fah_to_kel(x)
}
\arguments{
\item{x}{of Fahrenheit degrees is a numerical (double)}
}
\value{
Kelvin(K) as numerical (double)
}
\description{
fah_to_kel is a function to convert Fahrenheit degrees to Kelvin.
Insert temperature in Fahrenheit degrees to get the value in Kelvin.
For further information access:
https://github.com/marcelofukuoka/conversorR
}
\examples{
fah_to_kel(80)
}
