% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kel_to_cel.R
\name{kel_to_cel}
\alias{kel_to_cel}
\title{Convert Kelvin to Celsius degrees}
\usage{
kel_to_cel(x)
}
\arguments{
\item{x}{of Kelvin(K) is a numerical (double)}
}
\value{
Celsius degrees as numerical (double)
}
\description{
kel_to_cel is a function to convert Kelvin(K) to Celsius degrees
Insert temperature in Kelvin(K) to get the value in Celsius degrees.
For further information access:
https://github.com/marcelofukuoka/conversorR
}
\examples{
kel_to_cel(300)
}
