% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kel_to_fah.R
\name{kel_to_fah}
\alias{kel_to_fah}
\title{Convert Kelvin(K) to Fahrenheit degrees}
\usage{
kel_to_fah(x)
}
\arguments{
\item{x}{of Kelvin(K) is a numerical (double)}
}
\value{
Fahrenheit degrees as numerical (double)
}
\description{
kel_to_fah is a function to convert Kelvin(K) to Fahrenheit degrees
Insert temperature in Kelvin(K) to get the value in Fahrenheit degrees.
For further information access:
https://github.com/marcelofukuoka/conversorR
}
\examples{
kel_to_fah(299.82)
}
