% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit.R
\name{rfahrenheit}
\alias{rfahrenheit}
\title{Convert Celsius degrees to Fahrenheit}
\usage{
rfahrenheit(x)
}
\arguments{
\item{x}{of Celsius degrees is a numerical (double)}
}
\value{
Fahrenheit degrees as numerical (double)
}
\description{
rfahrenheit is a function to convert Celsius degrees to Fahrenhit.
Insert temperature in Celsius degrees to get the value in Fahrenheit degrees.
For further information access:
https://github.com/marcelofukuoka/conversorR
}
